Separate enum tests
