+++ tests/PHPStan/Reflection/Deprecation/data/CustomDeprecated.php
@@ -1,0 +1,15 @@
+<?php // lint >= 8.1
+
+namespace CustomDeprecations;
+
+#[\Attribute(\Attribute::TARGET_ALL)]
+class CustomDeprecated {
+
+	public ?string $description;
+
+	public function __construct(
+		?string $description = null
+	) {
+		$this->description = $description;
+	}
+}
