+++ tests/PHPStan/Reflection/Deprecation/data/deprecations.php
@@ -2,18 +2,6 @@
 
 namespace CustomDeprecations;
 
-#[\Attribute(\Attribute::TARGET_ALL)]
-class CustomDeprecated {
-
-	public ?string $description;
-
-	public function __construct(
-		?string $description = null
-	) {
-		$this->description = $description;
-	}
-}
-
 class NotDeprecatedClass
 {
 	const FOO = 'foo';
@@ -161,22 +149,3 @@ function doubleDeprecatedFunctionOnlyAttributeMessage() {}
 /** @deprecated phpdoc */
 #[CustomDeprecated()]
 function doubleDeprecatedFunctionOnlyPhpDocMessage() {}
-
-
-#[CustomDeprecated]
-enum MyDeprecatedEnum: string
-{
-	#[CustomDeprecated('custom')]
-	case CustomDeprecated = '1';
-
-	/**
-	 * @deprecated phpdoc
-	 */
-	case PhpDocDeprecated = '2';
-
-	#[\Deprecated('native')]
-	case NativeDeprecated = '3';
-
-	case NotDeprecated = '4';
-
-}
